chai tests: Beatmap, Beatmapset, Changelog
