+++ package.json
@@ -32,9 +32,11 @@
   ],
   "license": "Unlicense",
   "devDependencies": {
+    "@types/chai": "^5.0.1",
     "@types/node": "^22.13.5",
     "@types/ws": "^8.5.10",
     "ajv": "^8.17.1",
+    "chai": "^5.2.0",
     "dotenv": "^16.4.5",
     "ts-json-schema-generator": "^2.3.0",
     "typedoc": "^0.27.9",
